requirements.txt
requests
python-dotenv
numpy
faiss-cpu
tqdm
pandas